✅ 37. Locking Down DeleteItem Permissions (7min)

+++ backend/src/resolvers/Mutation.js
@@ -41,8 +41,16 @@ const Mutations = {
     async deleteItem(parent, args, ctx, info) {
         const where = { id: args.id };
         //find the item
-        const item = await ctx.db.query.item({where}, `{ id, title }`)
+        const item = await ctx.db.query.item({where}, `{ id title }`)
         //check if they own the item
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission =>
+            ['ADMIN', 'ITEMDELETE'].includes(permission)
+        );
+
+        if (!ownsItem && !hasPermissions) {
+            throw new Error("You don't have permission to do that!");
+        }
         //delete the item
         return ctx.db.mutation.deleteItem({where}, info);
     },
